02_Code/requirements.txt
geopandas
pandas
pystac
scikit-learn
requests
rasterio
shapely
tqdm
ultralytics
sahi
pyyaml